streamlit-app/requirements.txt
# Dépendances pour l'application Streamlit MEAL Nutrition
# CARIS Foundation International

# Core Streamlit
streamlit>=1.28.0

# Authentification
streamlit-authenticator>=0.3.0
PyYAML>=6.0
bcrypt>=4.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Visualizations
plotly>=5.15.0

# AI/LLM Integration
google-generativeai>=0.3.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Date/Time
python-dateutil>=2.8.0
